파이썬 코드 합침
